src/scraper/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
playwright>=1.40
pdfplumber>=0.10
psycopg2-binary>=2.9
python-dotenv>=1.0
fuzzywuzzy>=0.18
python-Levenshtein>=0.23
openpyxl>=3.1
pandas>=2.0
feedparser>=6.0
chromadb>=0.4
